Move SortableBookshelfReorderGrid and SortableBookshelfCard to widgets directory
